SIGNIFICANTLY restructure hook load order and default page content.
Make default main section content a primary column with a default loop and a secondary column with a default sidebar.

This has been implemented as intended (for now), but without docs or standard file organization (thanks to the rush for the US Chess build).

This requires docs updating and file reorganization.